app/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.0
h11==0.14.0
httptools==0.6.1
idna==3.10
mypy==1.11.2
mypy-extensions==1.0.0
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
